Add some more failing tests.
DIRSTACK and dirs disagree on their contents a bit.

dirs always shows the home directory as ~.
DIRSTACK expands ~ in all but the current/zeroeth element.

/home/user $ mkdir foo
/home/user $ pushd foo
~/foo ~
/home/user $ dirs
~/foo ~
/home/user $ declare -p DIRSTACK
declare -a DIRSTACK='([0]="~/foo" [1]="/home/deryni")'
/home/user $ pushd +1
~ ~/foo
/home/user $ declare -p DIRSTACK
declare -a DIRSTACK='([0]="~" [1]="/home/deryni/foo")'

So ~+0 and ~-0 fail to work currently as they leave the unexpanded ~ in the
result.